Fix for flood protection in ::Module::ChanOp.
There was a bug leading to timestamps not being checked correctly, so exceeding
the maximum number of messages, even across a very long timeframe, would lead to
mistaken kicks.

Refactored to make sure that only timestamps of messages seen within the desired
window are remembered, then check.

This patch is tested and in-use on a active IRC channel and behaves
appropriately.