Store passwords in bot store hashed, not plain.
This commit avoids storing user's passwords in plain text in the bot store.

OK, users should obviously not be using a valuable password (i.e. one used for
anything else) to identify to an IRC bot, but in my opinion, there's still no
reason to store plain text passwords instead of a hashed version.

This commit uses Crypt::SaltedHash to take care of generating secure salted
hashed versions of users passwords, and verifying them.  Existing plaintext
passwords will continue to be recognised, but when the user changes their
password, a hashed version of the new password will be stored.