Check for existing table using a '%' schema.
In create_table(), the check for an existing table is done with a call
to table_info. This call returns a statement handle that when fetched
only yields undef(), meaning create_table will always try to create a
new table, even if there is an existing one. This test works again by
adding a '%' schema filter to the table_info arguments.